add components at step0
