--- front/src/components/templates/RegisterCard.tsx
@@ -1,15 +1,0 @@
-import {FC} from "react"
-import { ReactNode } from 'react'
-
-interface IProps {
-name: string
-children: ReactNode
-}
-
-export const RegisterCard: FC<IProps> = ({ name }) => {
-  return (
-    
-            
-    
-  )
-}
