+++ front/src/pages/step0/SecCard.tsx
@@ -1,6 +1,9 @@
 import {FC} from "react"
 import { RegisterNextButton } from '../../components/parts/RegisterNextButton'
 import { GitHubButton } from '../../components/parts/GitHubButton'
+import { RegisterCardTemprate } from './RegisterCardTemprate'
+import { RegisterCardHead } from "./RegisterCardHead"
+import { RegisterCardBody} from "./RegisterCardBody"
 
 export const SecCard: FC =()=> {
     return (
@@ -8,24 +11,21 @@ export const SecCard: FC =()=> {
                             justify-center
                             p-4 ">
 
-                    {/* ----- 0 step ------ */}
-        <div className="block max-w-sm rounded-lg bg-white text-center shadow-lg mx-auto">
-                <div className="rounded-t-lg bg-purple py-5 text-3xl font-bold tracking-wider text-white">
-                    <p>参加者登録</p>
-                </div>
+            {/* ----- 0 step ------ */}
+        <RegisterCardTemprate>
+            <RegisterCardHead name="参加者登録" children/>
+                <RegisterCardBody>
 
-                <div className="px-12 py-4 ">
-                        <div className="w-60 px-2 py-2 my-10
-                                        text-xl">
-                            <p>SNS登録</p>
-                        </div>
+                    <div className="w-60 px-2 py-2 my-10
+                                    text-xl">
+                        <p>SNS登録</p>
+                    </div>
 
                     <GitHubButton path=""/>
                     <RegisterNextButton path="" name="Next"/>
 
-                </div>
-                
-            </div>
+                </RegisterCardBody>
+        </RegisterCardTemprate>
         </section>
     )
 }
